requirements.txt
fastapi
uvicorn
python-multipart
mysql-connector-python
python-dotenv
jinja2
bcrypt